LayoutTests/http/tests/clear-site-data/set-and-clear-cookies.https-expected.txt

CONSOLE MESSAGE: First party cookies before clear-site-data: test_first_party_cookie=1
CONSOLE MESSAGE: First party cookies after clear-site-data:
Tests that first and third-party partitioned cookies are accepted.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS successfullyParsed is true

TEST COMPLETE


--------
Frame: '<!--frame1-->'
--------


--------
Frame: '<!--frame2-->'
--------
Cookies are: test_cookie = 1

--------
Frame: '<!--frame3-->'
--------
Cookies are:

--------
Frame: '<!--frame4-->'
--------
